Add text to video with a txt file
Previous commit used, the text explicitly given in the shell script.
But that's not safe as adding things to main shell
script could break the script.

So now we can give a txt file with all the texts we want to display on video.

RULES TO FOLLOW IN .txt FILE
	*File should be named "text.txt"
	*File should be inside the script directory
	*Lines should be in order where you want them to appear in video
	*When you need to break a sentence you should not use return(EnterKey). 	Instead of using return you should put in "\n" character where you		need to	break the sentence

There is a new python script(text_breaking.py) to handle the txt file
	*This will break the original text file into several txt files(one for each line) and will align all texts to be in center.
	*And store breaked files inside the media directory(eg:100.txt,101.txt)
	*File names start with 100, this is because we need to add text to the
	video in the order they are given

	**These newly created txt files will be automatically deleted once they
	  are no longer needed