Change IMAGE_DURATION automatically
In the previous commit audio stops mid-way when the picture slideshow ends.
This is because user passes the IMAGE_DURATION and TRANSITION_DURATION.
Because of the -shortest option used , audio ends midway when the slideshow ends.

But in this, IMAGE_DURATION changes automatically according to audio duration to keep the whole audio.
That means pictures transition through out the entire audio.
Thereofore audio does not stop mid-way.

Also a new font is used here for more readability.