Add lyrics synchronization
In the previous commit we were only able to display texts uniformly across the video .
But in this, we can give a srt file with the lyrics to a song and the script will match
those lyrics to specified times and show them accordingly in the right timeframe in the video.

STUCTURE
	*srt file should be inside scripts directory and should be named "lyrics.srt"
	*Images should be in media folder and you can decide the order of the image
		appearance by renaming image files in order you want. Renaming from 100 is recommended.(eg: 100.jpg, 101.jpg, ...)
	*Audio file should be inside media folder and should be named "song.mp3"

There is a new python script: "caption_centered.py"
	* Parsing of the srt file will be done by the "caption_centered.py" script
	* This will parse the srt file and extract the time range for each line in the srt file
		and store them in the "media/lyrics/timing" directory. (Two files for each time-range)
	* And this will extract lyrics and will create a file for each time-range in "media/lyrics" directory

This can be used to create lyrics videos to songs. (srt file can be downloaded with youtube captions)